requirements.txt
fastmcp>=2.11.3
requests
python-dotenv
